% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCIA_plots_JR.R
\name{biological_comparison}
\alias{biological_comparison}
\title{Perform biological annotation-based comparison}
\usage{
biological_comparison(factorizations, path.database, pval.thr = 0.05)
}
\arguments{
\item{factorizations}{= already computed factorizations}

\item{path.database}{= path to a GMT annotation file}

\item{pval.thr}{= p-value threshold (default to 0.05)}

\item{genes}{a vector of gene names according to HUGO nomenclature}
}
\value{
selectivity Selectivity (fraction of significant annotations per all significant factors)

nonZeroFacs Number of unique factors that were significant at least once

total_pathways = Number of clinical annotations with at least one significant factor component
}
\description{
Runs fgsea for the input gene list
}
