% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mcia_lfs.R
\name{corr_heatmap_fvl_ggplot2}
\alias{corr_heatmap_fvl_ggplot2}
\title{Plot a correlation heatmap of global scores versus features using ggplot2}
\usage{
corr_heatmap_fvl_ggplot2(
  global_scores,
  feature_mat,
  feature_name = "Feature",
  select_features = NULL
)
}
\arguments{
\item{global_scores}{the MCIA factor object calculated from block_matrix}

\item{feature_mat}{the input matrix used to calculate the MCIA factors}

\item{feature_name}{the name of the input}

\item{select_features}{a vector of numbers to filter features}
}
\value{
the ggplot2 object
}
\description{
Plots a heatmap of MCIA factos versus other vectors such as
the original feature vectors (membership analysis) or versus tasks to 
investigate the predictive power of these factors to some other vector of data.
}
