% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blocks.R
\name{data_blocks}
\alias{data_blocks}
\title{NCI-60 Multi-Omics Data}
\description{
A  dataset of mesurements of 12,895 mRNA, 537 miRNA,
and 7,016 protein variables (columns) on 21 cancer cell lines (rows) from
the NCI-60 cancer cell line database.

Data originally from the supplementary materials of Meng et. al, 2016.
}
\references{
\url{https://github.com/aedin/NCI60Example}
}
\keyword{data}
\keyword{multi-omics}
