% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_colors}
\alias{get_colors}
\title{Assigning colors to different omics (default: color-blindness friendly)}
\usage{
get_colors(
  mcia_result,
  color_pal = scales::viridis_pal,
  color_pal_params = list(option = "D")
)
}
\arguments{
\item{mcia_result}{object returned from nipals_multiblock() function}

\item{color_func}{a function which returns color palettes (e.g. scales)}

\item{color_params}{list of parameters for the corresponding function}
}
\value{
List of omics with assigned colors
}
\description{
Creates a list of omics and associated colors for plotting
}
\examples{
colors_omics <- get_colors(mcia_results)
}
