% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_metadata_colors}
\alias{get_metadata_colors}
\title{Assigning colors to different values of a metadata column 
(default: color-blindness friendly)}
\usage{
get_metadata_colors(mcia_result, coloring, color_func = scales::viridis_pal)
}
\arguments{
\item{mcia_result}{object returned from nipals_multiblock() function}

\item{coloring}{an integer or string specifying the column that will be
used for coloring}

\item{color_func}{a function which returns color palettes}
}
\value{
List of metadata columns with assigned colors
}
\description{
Creates a list of metadata columns and associated colors
  for plotting
}
\examples{
colors_omics <- get_metadata_colors(mcia_results, "cancerType")
}
