% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TV.R
\name{get_TV}
\alias{get_TV}
\title{Computes the total variance of a multi-omics dataset}
\usage{
get_TV(ds)
}
\arguments{
\item{ds}{a list of multi-omics dataframes/matrices in "sample x variable" format}
}
\value{
the total variance of the dataset (i.e. sum of block variances)
}
\description{
Computes the total variances of all data blocks in a multi-omics dataset, 
intended for datasets that do not use `CCpreproc`
}
\examples{
tot_var <- get_TV(data_blocks)
}
