% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mcia_lfs.R
\name{global_loadings_heatmap_ComplexHeatmap}
\alias{global_loadings_heatmap_ComplexHeatmap}
\title{Plotting a heatmap of global_loadings versus features}
\usage{
global_loadings_heatmap_ComplexHeatmap(
  global_loadings,
  data_blocks,
  omic_name,
  select_features = NULL
)

global_loadings_heatmap_ComplexHeatmap(
  global_loadings,
  data_blocks,
  omic_name,
  select_features = NULL
)
}
\arguments{
\item{global_loadings}{the global_loadings matrix after running MCIA}

\item{data_blocks}{the list of matrices used to calculate the MCIA factors}

\item{omic_name}{the name of the omic that should be plot, should be in data_blocks}

\item{select_features}{a vector of numbers to filter features}
}
\value{
the ggplot2 object

the ggplot2 object
}
\description{
Plots a heatmap of MCIA global_loadings versus factors

Plots a heatmap of MCIA global_loadings versus factors
}
