% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_report.R
\name{gsea_report}
\alias{gsea_report}
\title{Perform biological annotation-based comparison}
\usage{
gsea_report(
  metagenes,
  path.database,
  factors = NULL,
  pval.thr = 0.05,
  nproc = 4
)
}
\arguments{
\item{metagenes}{Vector of gene scores where the row names are HUGO symbols}

\item{path.database}{path to a GMT annotation file}

\item{pval.thr}{= p-value threshold (default to 0.05)}

\item{factor}{vector of factors which should be analyzed}
}
\value{
data frame with the most significant p-value number of significant
pathways

the selectivity scores across the given factors
}
\description{
Runs fgsea for the input gene vector
}
