% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCIA_plots.R
\name{MCIA_plots}
\alias{MCIA_plots}
\title{MCIA_plots}
\usage{
MCIA_plots(
  mcia_result,
  plotType,
  orders = c(1, 2),
  clusters = "none",
  clusterColors = list("red", "green", "blue", "yellow", "brown", "orange", "maroon",
    "magenta", "turquoise", "darkgreen", "darkblue", "gold", "pink"),
  legend_loc = "bottomleft"
)
}
\arguments{
\item{mcia_result}{MCIA results object returned from `nipals_multiblock`}

\item{plotType}{Type of plot, with the following options \itemize{
\item `projection` - scatter plot of two orders of global and block scores (aka factors).
\item `projection_global` - scatter plot of two orders of global scores only (aka factors).
\item `gs_eigvals` - scree plot of global score eigenvalues.
\item `block_weights_heatmap` - a heatmap of block weightings for each order of global score.
}}

\item{orders}{Option to selecct orders of factors to plot against each other (for projection plots)}

\item{clusters}{Option to plot clusters in projection plots, with two options:\itemize{
\item `none` (default) - no clusters plotted
\item `metadata` - provide cluster labels in the `metadata` field of the `mcia_result` argument. 
`metadata` field must be a list of strings of cluster names for each sample/row in the dataset.  
}}

\item{clusterColors}{Option to specify cluster color strings.}

\item{legend_loc}{Option for legend location, or "none" for no legend.}
}
\description{
Function to generate descriptive plots for MCIA.
}
\details{
Plotting function for a variety of MCIA visualization options.
}
\examples{
# Plotting clusters with different colors
data(NCI60)
mcia_res <- nipals_multiblock(data_blocks,num_PCs = 10, plots = 'none', tol=1e-12)
CNS = 1:6; LEU = 7:12; ME = 13:21;
nameslist <- list()
nameslist[CNS] <- "CNS"
nameslist[LEU] <- "Leukemia"
nameslist[ME] <- "Melanoma"
mcia_res$metadata <- nameslist
clus_colors <- list("red", "green","blue")
MCIA_plots(mcia_res,'projection',orders = c(1,2), clusters = 'metadata', 
           clusterColors = clus_colors, legend_loc = "bottomleft")

}
