% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_NCI60.R
\name{metadata_NCI60}
\alias{metadata_NCI60}
\title{NCI-60 Multi-Omics Metadata}
\description{
Metadata for the included multi-omics dataset, denoting the cancer
type associated with each of the 21 cell lines.  

Data originally from the supplementary materials of Meng et. al, 2016.
}
\references{
\url{https://github.com/aedin/NCI60Example}
}
\keyword{data}
\keyword{multi-omics}
