% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nipalsMCIA-package.R
\docType{package}
\name{nipalsMCIA-package}
\alias{nipalsMCIA}
\alias{nipalsMCIA-package}
\title{nipalsMCIA: Multiple Co-Inertia Analysis via the NIPALS Method}
\description{
Computes Multiple Co-Inertia Analysis (MCIA), a dimensionality reduction (jDR) algorithm, for a multi-block dataset using a modification to the Nonlinear Iterative Partial Least Squares method (NIPALS) proposed in (Hanafi et. al, 2010). Allows multiple options for row- and table-level preprocessing, and speeds up computation of variance explained. Vignettes detail application to bulk- and single cell- multi-omics studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Muunraker/nipalsMCIA}
  \item Report bugs at \url{https://github.com/Muunraker/nipalsMCIA/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Mattessich \email{maximilian.mattessich@northwestern.edu} (\href{https://orcid.org/0000-0002-1233-1240}{ORCID})

Authors:
\itemize{
  \item Joaquin Reyna \email{joreyna@live.com} (\href{https://orcid.org/0000-0002-8468-2840}{ORCID})
  \item Edel Aron \email{edel.aron@yale.edu} (\href{https://orcid.org/0000-0002-8683-4772}{ORCID})
  \item Anna Konstorum \email{konstorum.anna@gmail.com} (\href{https://orcid.org/0000-0003-4024-2058}{ORCID})
}

}
\keyword{internal}
