% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmb_set_metadata.R
\name{nmb_set_metadata}
\alias{nmb_set_metadata}
\title{Function to set the metadata of a `NipalsResult` object.}
\usage{
nmb_set_metadata(nmb_object, nmb_metadata)
}
\arguments{
\item{nmb_object}{The `NipalsResult` object to set metadata for.}

\item{nmb_metadata}{A data frame of metadata for each sample in the dataset.
Each column must have length equal to the number of samples (rows)
in the dataset.}
}
\value{
A `NipalsResult` object with the provided metadata.
}
\description{
Function to set the metadata of a `NipalsResult` object.
}
\examples{
data("NCI60")
data_blocks_mae <- simple_mae(data_blocks,row_format="sample",
                              colData=metadata_NCI60)
mcia_out <- nipals_multiblock(data_blocks_mae, num_PCs = 10)
mcia_out <- nmb_set_metadata(mcia_out,metadata_NCI60)
}
