% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIPALS_MCIA.R
\name{omicadeInitialization}
\alias{omicadeInitialization}
\title{Omicade4 Initialization}
\usage{
omicadeInitialization(df)
}
\arguments{
\item{df}{the data frame to apply pre-processing to, in "sample" x "variable" format}
}
\value{
the processed data frame
}
\description{
Applies the variable and block level pre-processing as the 
"Omicade4" package (Meng et al. 2014)
}
\details{
Performs the following steps on a given data frame: \itemize{
\item Offsets data to make whole matrix non-negative
\item Divides each column by its sum 
\item Subtracts (row sum/total sum) from each row
\item Multiplies each column by sqrt(column sum/total sum)
\item Divides the whole data frame by its total variance (the sqrt of the sum of singular values)
}
}
