% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blocks.R
\name{data_blocks}
\alias{data_blocks}
\title{NCI-60 Multi-Omics Data}
\source{
Meng et. al, 2016 supplementary materials {https://doi.org/10.1093/bib/bbv108}
}
\value{
Large list with 3 elements (one for each omic)
}
\description{
A  dataset of measurements of 12,895 mRNA, 537 miRNA,
and 7,016 protein variables (columns) on 21 cancer cell lines (rows) from
the NCI-60 cancer cell line database.
}
\references{
\url{https://github.com/aedin/NCI60Example}
}
\keyword{data}
\keyword{multi-omics}
