% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_colors}
\alias{get_colors}
\title{Assigning colors to different omics (color-blindness friendly)}
\usage{
get_colors(mcia_out)
}
\arguments{
\item{mcia_out}{object returned from nipals_multiblock() function}
}
\value{
List of omics with assigned colors
}
\description{
Creates a list of omics and associated colors for plotting
}
\examples{
colors_omics <- get_colors(mcia_out)
}
