% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_scores_heatmap.R
\name{global_scores_heatmap}
\alias{global_scores_heatmap}
\title{Plotting a heatmap of global factors scores (sample v. factors)}
\usage{
global_scores_heatmap(mcia_result, coloring = NULL)
}
\arguments{
\item{mcia_result}{the mcia object matrix after running MCIA, must also
contain metadata with columns corresponding to coloring}

\item{coloring}{an integer or string specifying the column that will be
used for coloring}
}
\value{
ComplexHeatmap object
}
\description{
Plots a heatmap of MCIA global scores
}
