% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_scores_heatmap.R
\name{global_scores_heatmap}
\alias{global_scores_heatmap}
\title{Plotting a heatmap of global factors scores (sample v. factors)}
\usage{
global_scores_heatmap(
  mcia_results,
  color_col = NULL,
  color_pal = scales::viridis_pal,
  color_pal_params = list(option = "D"),
  separate_rects = TRUE
)
}
\arguments{
\item{mcia_results}{the mcia object matrix after running MCIA, must also
contain metadata with columns corresponding to color_col}

\item{color_col}{an integer or string specifying the column that will be
used for color_col}

\item{color_pal}{a list of colors or function which returns a list of colors}

\item{color_pal_params}{a list of parameters for the color function}

\item{separate_rects}{a boolean indicating whether or not to add white lines between the rectangles for visual clarity}
}
\value{
ComplexHeatmap object
}
\description{
Plots a heatmap of MCIA global scores
}
