% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_scores_heatmap.R
\name{global_scores_heatmap}
\alias{global_scores_heatmap}
\title{Plotting a heatmap of global factors scores (sample v. factors)}
\usage{
global_scores_heatmap(global_scores)
}
\arguments{
\item{global_scores}{the global_scores matrix after running MCIA}
}
\value{
the ggplot2 object
}
\description{
Plots a heatmap of MCIA global scores
}
