% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCIA_plots.R
\name{MCIA_plots}
\alias{MCIA_plots}
\title{MCIA_plots}
\usage{
MCIA_plots(mcia_result, plotType, orders = c(1, 2), clusters = "none")
}
\arguments{
\item{mcia_result}{MCIA results object returned from `nipals_multiblock`}

\item{plotType}{Type of plot, with the following options \itemize{
\item `projection` - scatter plot of two orders of global and block scores (aka factors)
\item `projection_global` - scatter plot of two orders of global scores only (aka factors).
\item `gs_eigvals` - scree plot of global score eigenvalues
\item `block_weights_heatmap` - a heatmap of block weightings for each order of global score
}}

\item{orders}{Option to selecct orders of factors to plot against each other (for projection plots)}

\item{clusters}{Option for list of indices of known clusters (for projection plots).}
}
\description{
Function to generate descriptive plots from MCIA
}
\details{
Plotting function for a variety of MCIA visualization options.
}
