% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCIA_plots.R
\name{MCIA_plots}
\alias{MCIA_plots}
\title{MCIA_plots}
\usage{
MCIA_plots(
  mcia_result,
  plotType,
  orders = c(1, 2),
  coloring = "none",
  labelColors = list("red", "green", "blue", "yellow", "brown", "orange", "maroon",
    "magenta", "turquoise", "darkgreen", "darkblue", "gold", "pink"),
  legend_loc = "bottomleft"
)
}
\arguments{
\item{mcia_result}{MCIA results object returned from `nipals_multiblock`}

\item{plotType}{Type of plot, with the following options \itemize{
\item `projection` - scatter plot of two orders of global and block scores (aka factors).
\item `projection_global` - scatter plot of two orders of global scores only (aka factors).
\item `gs_eigvals` - scree plot of global score eigenvalues.
\item `block_weights_heatmap` - a heatmap of block weightings for each order of global score.
}}

\item{orders}{Option to selecct orders of factors to plot against each other (for projection plots)}

\item{coloring}{Option to plot clusters/colors in projection plots, with two options:\itemize{
\item `none` (Default) - no clusters/colors plotted.
\item A character string of the column name of the `mcia_result$metadata` dataframe 
determining which color groupings to use (projection plots only) 
}}

\item{labelColors}{Option to specify  colors for labels.}

\item{legend_loc}{Option for legend location, or "none" for no legend.}
}
\value{
Displays the desired plots
}
\description{
Function to generate descriptive plots for MCIA.
}
\details{
Plotting function for a variety of MCIA visualization options.
}
\examples{
data(NCI60)
mcia_res <- nipals_multiblock(data_blocks, metadata = metadata_NCI60, num_PCs = 10, 
                              plots = 'none', tol=1e-12)
clus_colors <- list("red", "green","blue")
MCIA_plots(mcia_res,'projection',orders = c(1,2), coloring="cancerType",
           labelColors = clus_colors, legend_loc = "bottomleft")
}
