% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_NCI60.R
\name{metadata_NCI60}
\alias{metadata_NCI60}
\title{NCI-60 Multi-Omics Metadata}
\source{
Meng et. al, 2016 supplementary materials {https://doi.org/10.1093/bib/bbv108}
}
\value{
List with 21 elements
}
\description{
Metadata for the included multi-omics dataset, denoting
the cancer type associated with each of the 21 cell lines.
}
\references{
\url{https://github.com/aedin/NCI60Example}
}
\keyword{data}
\keyword{multi-omics}
