% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gs.R
\name{predict_gs}
\alias{predict_gs}
\title{Prediction of new global scores based on block loadings and weights}
\usage{
predict_gs(bl, bw, df)
}
\arguments{
\item{bl}{a list of matrices of block loadings, where each matrix corresponds
to one omics type with dimensions "features" x "number of loadings"}

\item{bw}{a matrix of block weights, with dimensions "number of omics" x "number of loadings"}

\item{df}{a list of data matrices to make predictions from, where each entry 
corresponds to one omics type in "sample" x " features" format.
Feature and omic order must match `bl`. Pre-processing should also match the data
used to generate `bl` and `bw`.}
}
\value{
a matrix of predicted global scores, in form
}
\description{
Uses previously-computed block scores and weights to compute 
a global score for new data.
}
\details{
Projects the new observations onto each block loadings vector, then 
weights the projection according to the corresponding block weights.
}
\examples{
deflated_data <- deflate_block_bl(data_frame,block_loading)

}
