% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_plot.R
\name{projection_plot}
\alias{projection_plot}
\title{projection_plot}
\usage{
projection_plot(
  mcia_results,
  plotType,
  orders = c(1, 2),
  color_col = NULL,
  color_pal = scales::viridis_pal,
  color_pal_params = list(option = "D"),
  legend_loc = "bottomleft"
)
}
\arguments{
\item{mcia_results}{MCIA results object returned from `nipals_multiblock`}

\item{plotType}{Type of plot, with the following options \itemize{
\item `projection` - scatter plot of two orders of global and block scores (aka factors).
\item `projection_global` - scatter plot of two orders of global scores only (aka factors).
}}

\item{orders}{Option to select orders of factors to plot against each other (for projection plots)}

\item{color_col}{an integer or string specifying the column that will be
used for color_col}

\item{color_pal}{a list of colors or function which returns a list of colors}

\item{color_pal_params}{a list of parameters for the color function}

\item{legend_loc}{Option for legend location, or "none" for no legend.}
}
\value{
Displays the desired plots
}
\description{
Function to generate a projection plot of MCIA results.
}
\details{
Plotting function for a projection plot.
}
\examples{
data(NCI60)
mcia_results <- nipals_multiblock(data_blocks, metadata = metadata_NCI60,
                              num_PCs = 10, plots = "none", tol=1e-12)
projection_plot(mcia_results, "projection", orders = c(1,2),
  color_col = "cancerType", legend_loc = "bottomright")
}
