% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_plot.R
\name{projection_plot}
\alias{projection_plot}
\title{projection_plot}
\usage{
projection_plot(
  mcia_result,
  plotType,
  orders = c(1, 2),
  color_col = NULL,
  color_func = scales::viridis_pal,
  color_params = list(option = "D"),
  legend_loc = "bottomleft"
)
}
\arguments{
\item{plotType}{Type of plot, with the following options \itemize{
\item `projection` - scatter plot of two orders of global and block scores (aka factors).
\item `projection_global` - scatter plot of two orders of global scores only (aka factors).
}}

\item{orders}{Option to select orders of factors to plot against each other (for projection plots)}

\item{color_col}{Option to plot clusters/colors in projection plots, with two options:\itemize{
\item `none` (Default) - no clusters/colors plotted.
\item A character string of the column name of the `mcia_result$metadata` dataframe 
determining which color groupings to use (projection plots only) 
}}

\item{color_func}{a function which returns color palettes (e.g. scales)}

\item{color_params}{list of parameters for the corresponding function}

\item{legend_loc}{Option for legend location, or "none" for no legend.}

\item{mcia_results}{MCIA results object returned from `nipals_multiblock`}
}
\value{
Displays the desired plots
}
\description{
Function to generate a projection plot of MCIA results.
}
\details{
Plotting function for a projection plot.
}
\examples{
data(NCI60)
mcia_results <- nipals_multiblock(data_blocks, metadata = metadata_NCI60,
                              num_PCs = 10, plots = "none", tol=1e-12)
projection_plot(mcia_results, "projection", orders = c(1,2),
  color_col = "cancerType", legend_loc = "bottomright")
}
