% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_load_ord.R
\name{vis_load_ord}
\alias{vis_load_ord}
\title{Visualize ranked loadings}
\usage{
vis_load_ord(gl_f_ord, colors_omics, n_feat = 15)
}
\arguments{
\item{gl_f_ord}{Ranked loading dataframe output from ord_loadings() function}

\item{colors_omics}{named list of colors associated with omics,
output of get_colors() function}

\item{n_feat}{number of features to visualize}
}
\description{
Visualize a scree plot of loadings recovered from 
nipalsMCIA() output loadings matrix ranked using the 
ord_loadings() functions
}
\examples{
vis_load_ord(all_pos_1_df, colors_omics = colors_omics)
}
