% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_load_plot.R
\name{vis_load_plot}
\alias{vis_load_plot}
\title{Visualize all loadings on two factor axes}
\usage{
vis_load_plot(mcia_out, axes = c(1, 2), colors_omics)
}
\arguments{
\item{mcia_out}{object returned from nipals_multiblock() function}

\item{axes}{list of two numbers associated with two factors to visualize}

\item{colors_omics}{named list of colors associated with omics,
output of get_colors() function}
}
\value{
Plot of MCIA feature loadings for chosen axes
}
\description{
Visualize all loadings recovered from 
nipalsMCIA() output loadings matrix ranked using across two factor axes
}
\examples{
vis_load_plot(mcia_results, axes = c(1, 4), colors_omics)
}
